Refactored sales code into sub modules.
